backend/requirements.txt
fastapi
uvicorn
google-genai
openfoodfacts
pydantic
requests
uvicorn